requirements.txt
aiogram==2.7
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Migrate==4.0.1
Flask-SQLAlchemy==3.0.2
frozenlist==1.3.3
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.1
multidict==6.0.4
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
SQLAlchemy==1.4.46
urllib3==1.26.13
Werkzeug==2.2.2
yarl==1.8.2
